+++ src/main/kotlin/nn/activation/Signum.kt
@@ -2,11 +2,11 @@ package nn.activation
 
 class Signum: Activation()
 {
-    override fun forward(input: Float): Float
+    override fun forward(input: Double): Double
     {
-        return if (input > 0) 1.0f else 0.0f
+        return if (input > 0) 1.0 else 0.0
     }
-    override fun backward(input: Float): Float
+    override fun backward(input: Double): Double
     {
         throw NotImplementedError()
     }
