+++ src/main/kotlin/nn/learn/HebbLearner.kt
@@ -1,5 +1,6 @@
 package nn.learn
 
+import koma.matrix.ejml.backend.times
 import nn.DataVector
 import nn.Net
 
@@ -10,7 +11,7 @@ class HebbLearner(private val net: Net,
     {
         val layer = this.net.layers[0]
         val output = layer.activation.forward(layer.forward(input))
-        val diff = (label - output) * this.learningRate
+        val diff = (label - output) * this.learningRate.toDouble()
         val deltas = diff * input
 
         layer.weights += deltas
